requirements.txt
antiorm==1.2.1
cffi==1.15.0
cryptography==3.0
db==0.1.1
db-sqlite3==0.0.1
pycparser==2.21
PyYAML==6.0
six==1.16.0
